Exposing internals for unit testing breaks the "lightweight" concept
I initially exposed this to unit test the custom "extend" implementation
without jquery, but the test named "passing options reference" already
test if options are being modified when removing the cookie. There is no
need to duplicate the test here.

Also, an exposed internal implementation might be used in the wild, so
let's avoid that.