This test is unnecessary
It was initially created to confirm "removeCookie" was returning "false"
in the case it couldn't be able to remove the cookie. But the only thing
it is doing is to stub the API and force returning false, there is no
real value on this test.

If you change the boolean return of "removeCookie", 2 tests fail anyway,
"removeCookie: when cookie does not exist" and "removeCookie: when
sucessfully deleted". All evidences points out this test is really
unnecessary and non-deterministic.

Also, this test relies in internal behavior for stubbing, future code
changes might require the modification of an unrelated test, which is
not ideal.