Enhances README documentation
Remove "Cookie Options" concept and instead consistently use the "Cookie
Attributes" concept. My theory is that people think that "options" are
customized to the more convenient defaults, while attributes make it
clear they are just passed through to the cookie attributes in the
browser.

Add the default browser behavior in bold right below each attribute,
along with an example of how to create, read and remove.