I believe this fix was mistaken
In the issue carhartl/jquery-cookie#155, maxim-dev says "Of course, I
had to code the name using urlencode() (for PHP, for example), but I did
not (my fault, I admit it)".

So, the user acknowledged that he was wrong and the issue was about the
wrong "$.removeCookie" return value instead of the fact the key had the
"[]" square bracket characters.

In RFC 6265 it's clear that a token (cookie-name) should NOT contain
square brackets characters. In the issue above, the brackets in the
cookie-name was the problem in the first place, and the OP acknowledged
that, so the fix to allow square brackets in the name is unnecessary.