Release on npm as the last step
IF we do this, then the "view documentation for the latest release" link
in the README shows on npm the current version instead of the previous.

That link is not even needed on npm. We just keep it there because there
is no procedure to remove it when a new version is released.