Make expires test agnostic to DST change
The test setup itself was assuming a different implementation. While
`setDate()` will respect values causing the new date to overlap a DST
change, e.g. adapt the hour accordingly, the underlying implementation
does not. Therefore we need to set up the expected date to assert for
differently, in line with the implementation that is.