Fix false positive in missing semicolon test
First of all including the utils script had been accidentally removed
earlier.

But attaching an `error` listener to an iframe is no longer possible due
to origin restrictions (yielding a generic "Script Error"), at least
when running the test in the headless chromium.

For this reason we didn't notice a. the missing utils file and b. the
problem with the missing semicolon, which the test is supposed to
verify.

The inspected `loadedSuccessfully` flag was never changed and always
(unexpectedly) true.

Also note that the problem only occurs with the non-minified umd
distribution variant.