Expose default path via api
So far we had been applying the default `path: '/'` as sort of a
"hidden" default, and it wasn't exposed in the defaults object
accessible as `Cookies.defaults`. This is rather unexpected and we start
exposing the path as part of the defaults to avoid surprises. This
change will also make it slightly more intuitive to change the behavior
to a browser's default behavior where a cookie is only valid at the
current page: instead of having to specify `path: ''` in the defaults we
can now do `delete Cookies.defaults.path`. The former is rather
obfuscating the intent, and a leaking abstraction.

Fixes #541