Support case insensitive url encoding
Lowercase escapes are legal, `decodeURIComponent()` supports them, and
we should too.

The most difficult part is to know when we are dealing with hexadecimal
pairs and when not. Supporting only uppercase characters made
interpreting strings slightly more constrained, nevertheless even in
this case we might run into a scenario where we encounter strings that
look like encoded characters but aren't, for instance "%A1".

The solution to this is to no longer permit strings with mixed unencoded
and encoded characters, which seems like a rather rare edge case to
support. Either a cookie value is a fully url encoded string (which is
ensured when we're writing the cookie via our own api) or we treat it as
non-readable otherwise. In such fully encoded strings "%" would appear
as "%25", e.g. "%A1" would look like "%25A1". In case someone needs to
work with such mixed encoding cookie values, they would need to resort
to implementing their own specialized converter.

Closing #466