Change `withConverter()` to require explicit type
So far when passing a function as argument to `withConverter()` it was
turned into a read converter. Setting up a write converter required to
pass options with a `write` property providing the desired converter.

This simplifies the api to always require the type (read/write) of a
converter to be specified explicitly, there should be only one way to
set up one or both converters.

Closes #570