Simplify read conversion behavior
Removing subtle fallback to the standard, internal converter by
returning a falsy value in the custom read converter. Exposing the
`rfc6265Converter` converter as part of the api instance allows users to
write self-contained converters without relying on internals.

So far this behavior was required to allow for custom converters that
only act on specific keys (and otherwise forwards the value to the
internal, default conversion). Such a converter can now be implemented
like so:

```
var customReadConverter = (value, key) => {
  if (value === 'special') {
    return value.toUpperCase()
  }
  return Cookies.rfc6265Converter.read(value)
}
```

Closes #571