Remove `defaults` in favor of builder
We start encapsulating the default cookie attributes. As a consequence
these defaults can no longer change per instance of an api over the
course of its lifetime. This is desired (more predictable behavior as
global defaults can no longer be changed in-between), and also, this way
it's now consistent with how converters are set up.

The standard api instance remains the same and comes with the default
'/' for path and the standard (rfc6265) converter.

Setting up converters and default attributes can be chained, multiple
times.

Closes #572