Add possibility for people to donate crypto
I was chatting with @carhartl and there's a significant amount of traffic coming from a lot of different places to js-cookie Github repo. Likewise, I've read countless posts talking about how many Open Source devs work below the poverty line.

We decided to try out and put a couple of public addresses on the main page to see if people who use js-cookie would be keen to donate to fund other Open Source projects.

I wasn't able to make significant contributions to Open Source in the past few years, that's my motivation. The idea is to use js-cookie visibility and the positive price predictions of BTC and ETH to make that stash grow in a way that creates a significant help to the community in the near future. It's the least I can do with my current time available.