Fix problem running browserstack tests on M1
The browser-stack runner was failing to verify the binary and attempted
to download a binary from a non-existent AWS S3 location (there's no
specific binary for M1).

Version 0.9.5 of the browserstack-runner fixed this, but wasn't
available on npm, therefore we're referencing the version via GitHub
commit.

Note it's still not possible to run browserstack tests just so on M1 - a
manual download plus setting up the `BROWSERSTACK_LOCAL_BINARY_PATH` env
var accordingly is still required to avoid attempting a broken download.