Ensure consistency w/ get('name') + get() behavior
In the case of two (or more) cookies with the same name visible to the
current scope, so far we've been retrieving the first one we find for
`get('name')` and the last one for `get()` (because in the latter case
we continued the loop and eventually overwrote a duplicate cookie name
in the returned result/object).

This change makes the behavior at least consistent, so there aren't
different cookies in the result for both calls.

We should be returning all visible cookies though, but the API change
for this is still being discussed.. see:
https://github.com/js-cookie/js-cookie/issues/813

Also see: https://github.com/js-cookie/js-cookie/wiki/Cookies-with-duplicated-name